// frontend/src/generated/model/practiceDirection.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Polyglot V2 API
 * OpenAPI spec version: 0.1.0
 */

export type PracticeDirection =
  (typeof PracticeDirection)[keyof typeof PracticeDirection];

export const PracticeDirection = {
  target_to_support: "target_to_support",
  support_to_target: "support_to_target",
  bidirectional: "bidirectional",
} as const;