// frontend/src/generated/model/practiceMode.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Polyglot V2 API
 * OpenAPI spec version: 0.1.0
 */

export type PracticeMode = (typeof PracticeMode)[keyof typeof PracticeMode];

export const PracticeMode = {
  cards: "cards",
  recognition: "recognition",
  recall: "recall",
  mixed: "mixed",
} as const;