// frontend/src/generated/model/practiceStackResponse.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Polyglot V2 API
 * OpenAPI spec version: 0.1.0
 */
import type { PracticeStackKind } from "./practiceStackKind";
import type { StackMemberModel } from "./stackMemberModel";

export interface PracticeStackResponse {
  checksum: string;
  created_at: string;
  language_pack_revision_id: string;
  members: StackMemberModel[];
  name: string;
  pedagogical_day: string | null;
  profile_id: string;
  source_refs: string[];
  stack_id: string;
  stack_kind: PracticeStackKind;
}